backend/requirements.txt
Flask==3.1.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Security==5.6.2
Flask-Mail==0.10.0
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.0.1
python-docx==0.8.11
openpyxl==3.1.2
PyPDF2==3.0.1
reportlab==4.0.4
Pillow==10.0.1
requests==2.31.0
python-multipart==0.0.6
Werkzeug==3.1.3
pdfplumber==0.9.0
python-dotenv==1.0.0
